cnn module update prototype
working prototype without docs/tests. This commit significantly modifies 
the api for model training and prediction with cnn's. In particular, a 
single preprocessor belongs to a model object, and that preprocessor is 
used during training, validation, and prediction. The preprocessor 
handles both file-by-file training/prediction and clips of longer files.