change Type of BaseAction.params from dict to pd.Series
This change leaves the api almost identical, but allows a nice syntax 
for accessing/setting parameters or even just auto-completing to see 
what parameters are part of the params object, e.g.a 
`action.params.my_param = 3`. The action.set() and action.get() still 
work, as does loading args from a function and making sure set does not 
provide unexpected args. However, as with pd.Series, the user can 
accidentally set action.params.not_a_parameter = 0 for a parameter that 
does not exist. The use of auto-complete to check params keys should 
help  with this in practice.