support writing metadata for audio files Fixes #391
I added support for writing audio metadata in the Audio class by relying on soundfile.SoundFile. By default, the Audio.save function now first writes a file, then opens the file with soundfile.SoundFile and updates the metadata using __setattr__ for each field that is (1) in the list of supported fields and (2) has a value other than None in Audio object's .metadata dictionary.