add audio_add_noise in `actions`
audio_add_noise generates a noise of chosen color and dB level, and mixes it with the original signal (with chosen dB applied). The signal and noise dB levels can be (min,max) to randmoly select from a range.

also round n_sample to int in Audio.noise to resolve bug with floating point n samples