Change default args from lists to tuple
Default arg as a (mutable) list can cause unexpected/surprising behavior because the list is only initialized once and can be modified. I've changed any default arguments that were lists to tuples. Behavior is the same, but note that a 1-element tuple must be written with a comma, eg (0,) because (0) is not interpreted as a tuple