workaround for adding trace with Spectrogram
Spectrogram's setattr raising AttributeError means that copy.deepcopy() will fail. As a workaround, if AttributeError is raised when trying to copy sample.data in to saple.trace[], it assigns the original object instead of copying (if immutable this isnt an issue). However, this isn't a good solution, and change immutable class implementation when #671 is addressed it should be changed