change default resample type to 'soxr_hq'
previoius default was 'kaiser_fast'. This change mirrors a change of default resampling in librosa v0.10.0. Resovles #674 [Dependency] `resampy` #674

also require librosa >=0.10.0

this corresponds to the switch in Librosa from 'kaiser_best' resample default (and default support) to 'soxr_hq' resample default. In the new version 'kaiser_best' requires the user to install resampy, and in the old version 'soxr_hq' requires the user to install soxr, so its best that we require the newer librosa version to avoid missing dependencies.