warn usr if load_model loads model of old version
load_model loading an object of a previous opensoundscape version will not work properly because its an entire pickled object. Instead we need to save a pickled dictionary with save_torch_dict and load_torch_dict. load_model now warns the user if load_model() loads an object with a value of .opensoundscape_version that doesnt match the current version.