raise GetDurationExceptions in make_clip_df
make_clip_df now raises a custom exception class if raise_exceptions=True (default of False retains previous behavior of adding a row for the file with nan for start and end times). This raise_exceptions=True option is used by BoxedAnnotations.one_hot_clip_labels() and gives a helpful error message to the user explaining that they used full_duration=None, and the function failed to get the durations of audio files.

This commit also adds tests for BoxedAnnotations and