update cnn tutorial: train from raven annots
Added both data prep from Raven annotations and training a CNN on these annotations (both from Louis' tutorials for opso manuscript) to the cnn training tutorial notebook. However, the notebook would take forever to run and requires downloading a long file. I limited it to one epoch, but still we need to figure out what we want to do in the tutorial vs leave elsewhere.

I'm committing this version so that we can share it with anyone interested in seeing the in-progress version