refactor audio.extend
now two separate methods: audio.extend_to(duration) achieves desired duration, and allows duration<self.duration (trims) or >self.duration (extends with silence). audio.extend_by(duration) adds silence to the end, and does not allow duration<0.

The reason for this is that the expected behaviour of Audio.extend() is ambiguous, and also it was raising an error if duration<self.duration which caused unexpected errors.

resolves Audio.extend raises error if length < duration #695