resolve #792: no default shape of SpecPre
changes default shape of SpectrogramPreprocessor to None, retaining the original shape of the spectrogram. However, CNN class still has default sample shape of (224,224,3) which is passed when creating the preprocessor and will continue to be the default sample shape when using the CNN class.

I also refactored the preprocessor so that the shape is given as three separate arguments (height, width, channels) for clarity instead of one. I changed the attributes added to AudioSamples to match these (.height, .width, .channels) and refactored relevant code.

You can now specify None for just height or width while providing the other value, to retain just one original dimension of the spectrogram.