dont log_graph in wandb.watch
log_graph is causing errors when trying to continue training a model. Maybe this is a wandb bug, not sure. I get the error message "ValueError: You can only call `wandb.watch` once per model.  Pass a new instance of the model if you need to call wandb.watch again in your code." This happens both if I save -> load or even just run .train() twice for the same object.

I'm just going to use False as the default value and expose it as a configuration option in cnn.wandb_logging['log_graph']