resolves #784 BA.one_hot_clip_labels bug
BoxedAnnotations() doesn't create one-hot labels DataFrame for empty Raven selections file #784

The bug had to do with how we choose which audio files to create one-hot-clip-labels for. Since it was using the table of annotations, it would skip any audio files without annotations. I changed it to use an attribute of the BoxedAnnotations object .audio_files, which is a list of audio file paths. The user can also specify the list of audio files to create boxedannotations for. This also led to a refactor of .to_raven_files() along similar lines. Note that the to_raven_files function relies on exact matches between the list of audio_files and the 'audio_file' column of annotations in .df