refactor safe_dataset resolve #672
"No safe samples" exception in SafeDataset should include traceback of (last?) error #672

now _safe_get_item returns the sample or the exception raised, rather than the sample or None. This makes it easier to debug because you can inspect or raise the exc, or raise a new exception from the exc.

In __getitem__, if all samples fail to load the raised exception is rased _from_ the most recent preprocessing exception so that the user can inspect the preprocessing exception