resolves #793 dont clip spec values
Refactored Spectrogram class to not clip values when initialized (used to clip to decibel_limits by default). Instead, .plot() and .to_image() functions now have a range parameter specifying an input range (min,max) to clip to.

In order to preserve behavior / output scores of the ribbit function, I added an argument and a line to ribbit() that clips values of the spectrogram to a range (default (-100,-20)) before the rest of the algorithm. This results in creating the same ribbit scores as previous versions.