initial working pt lightning refactor
here, I've implemented Lightning class analogous to previous CNN (many of same functinos) as a subclass of LightningModule. But, I think we might actually want to wrap LightningModule in another class. One reason is that LightningModule.train() is used as a "set_train_mode" functionality, so we can't use train() like we did in CNN. We could use a different name (here train_me()), but it may end up being better to wrap the OPSOLightningModule in another class CNN()/AudioModel/etc. This would avoid the awkwardness of the train_me() function initializing a lightning.Trainer with module=self - I don't know if that's bad or fine but it suggests the opso API goes one level higher and wraps LightningModule

missing some essentials like model saving/loading, and getting device mismatches on mps, but generally working