Added _required_cols and _standard_cols as class variables for BoxedAnnotations.
This is useful for not having to hardcode these in different methods. I did not, however, change the defintion of `standard_columns` in `from_raven_files(...)` since this was different from `_standard_cols` in the `__init__` function. I wasn't sure if I should change it or not and didn't want to cause unexpected behaviours.