refactor configure_optimizers to fix continued training behavior
now the reset_optimizer and restart_scheduler flags are implemented (they were unused arguments of train() before), and are passed to configure_optimizers. configure_optimizers reloads state dicts from self.optimizer and self.scheduler if they're not none and not specified as reset/restart=True.

there was an issue with saving / copy.deepcopy pickling, which was due to unnecessarily including d['params'] in scheduler.load_state_dict(d) - it doesn't seem like params is supposed to be in the state dict for torch scheduler at all.

adds tests for correct resume or reset/restart behavior of optimizer and scheduler