add switch for embed() to return dfs or arrays
we need to return arrays rather than dfs if we have >2d result, for instance if we preserve spatial dimension. However, returning df is a useful default for scenarios where global avgpool is enabled and we just have a 1d feature per sample. So embed() now takes return_dfs flag, if False returns np array but default is True. If avgpool is set to False, overrides return_dfs to be False so that we can return >2d output.