use speed_of_sound attribute in methods
the SpatialEvent._localize_after_cross_correlation and SyncrhonizedRecorderArray.create_candidate_events methods were not using the objects' speed_of_sound attribute.

I also removed the default speed_of_sound value for methods, only retaining it as a default for initializing the two classes.

Now I think it's sensible to retain the global variable SPEED_OF_SOUND, which provides a default value for the initialization of these two classes. Otherwise we would just be hard-coding that default (in multiple places) or not having a default.

speed_of_sound is now a required argument for localize() and the associated functions specific to each algorithm

resolves Make SPEED_OF_SOUND a class attribute instead of module/global variable #725