make single_target a property w/setter
implement setter that changes loss class and torch metrics when setting self.single_target=True or False for SpectrogramModule class and subclasses.

The idea is that we can keep the metrics and loss function consistent with the single_target property: if the user changes single_target to False, loss and metrics should also change to match.

This could be a bit confusing if the user manually sets the loss_fn or changes the metrics, so it logs a message noting that it is updating the loss fn and torch metrics. Also, we recommend overriding _init_torch_metrics() method rather than directly modifying .torch_metrics, and consistent with this, the single_target setter calls _init_torch_metrics() if the value of self._single_target is being changed.