refactor for modular localization of events
creates a `localize_events_parallel` function and moves most of the input validation in localize_detections to create_candidate_events() where it makes sense for it to be. Now, localize_detections() just wraps create_candidate_events() with localize_events_parallel() and filters the returned events into localized and unlocalized lists.

resolves SynchronizedRecorderArray API should allow localizing a list of already-created events #1041

instead of adding a method to SynchronizedRecorderArray, it made more sense to add a function to the spatial_event module: this function is not dependent on anything within the SynchronizedRecorderArray object, it works directly on a list of SpatialEvent objects