method to extract clips aligned using tdoa
resolves add tools for reviewing localized SpatialEvents #981

the full review process should be provided in a notebook, but this may be the only functionality we want to include within opso for now: loads audio clips used to localize a PositionEstimate with offsets based on tdoa, such that the signal should align in time if tdoas are correct. Provides example for visualizing, and test.