enable automatic mixed precision on mps
"amp" will now be used on mps if model.use_amp=True. Previously, this raised an issue with mps device type (Apple silicon) but this was resolved in Pytoch 2.15.0
https://github.com/pytorch/pytorch/issues/88415

adds tests, separating tests for amp on cpu, cuda, and mps. Note that mps and cuda tests only run if the hardware is "available" on the testing machine