refactor to isolate shared code
this is a pretty big refactor aimed at making things like the BMZ Perch2/Birdnet/TF models and ONNXModel only need to implement batch_forward() and otherwise be able to use SpectrogramPreprocessor methods like .predict() and .embed()

The strategy is that you implement batch_forward as appropraite for the class, and it returns a dictionary of outputs. Then predict() and embed() just need to make the dataloader with self.predict_dataloader(), iterate the dataloader to get batches, call batch_forward() on each batch, and aggregate the results across batches.

Some things in the tutorials are likely broken after this refactor, I haven't run/tested/checked them.